refactor(catalog): extract useCatalog and useIsbnScanner hooks
Apply Separation of Concerns to CatalogScreen.tsx (804 → 261 lines).

- Add web-dashboard/src/hooks/useCatalog.ts
  Owns all book data-fetching, pagination, and CRUD state/logic.
  Exports UseCatalogReturn typed interface & shared pure helpers
  (getStatusFlags, getBookIcon, getPaginationWindow, STATUS_FILTERS).

- Add web-dashboard/src/hooks/useIsbnScanner.ts
  Owns camera BarcodeDetector lifecycle, scan loop, and debounced
  Open Library ISBN metadata lookup.
  VIBE Correction: AI suggested onScan: (isbn) => void (sync).
  Changed to onIsbnScanned: (isbn) => Promise<void> because the
  lookup is async and the void signature cannot propagate errors.

- Modify web-dashboard/src/screens/CatalogScreen.tsx
  Now pure JSX composition: const catalog = useCatalog(); +
  const scanner = useIsbnScanner(...); – no business logic inline.

Lab 3 – AI-Assisted Refactoring | Branch: refactor/catalog-custom-hook
